Refactor forgotten password related actions into separate contorller.
