+++ app/views/forgot_passwords/confirm_password_reset.html.haml
