+++ app/views/forgot_passwords/forgot_password.haml
