+++ app/views/forgot_passwords/new_password.html.haml
@@ -1,13 +1,15 @@
 %section.reset_password.container
   .row
     .col-sm-10.col-sm-offset-1
-      = form_tag set_password_path
+      = form_tag set_password_path do
         %header
           %h1 Reset Your Password
         %fieldset.form-group
           = label_tag 'New Password'
           .row
             .col-sm-4
-              = password_field_tag "password", "", class: 'form-control(type="password")'
+              = password_field_tag "password", "", class: 'form-control'
+        %fieldset.form-group
+          = hidden_field_tag :token, params[:token]
         %fieldset.form-group.action
           = submit_tag "Reset Password", class: "btn btn-default"
