+++ config/routes.rb
@@ -29,9 +29,9 @@ Myflix::Application.routes.draw do
   get 'register', to: 'users#new'
   post 'log_in', to: 'sessions#create'
   root 'pages#front'
-  get '/forgot_password', to: 'users#forgot_password'
-  get '/confirm_password_reset', to: 'users#confirm_password_reset'
-  get '/new_password/:token', to: 'users#new_password'
-  post '/password_reset', to: 'users#password_reset'
-  post '/set_password', to: 'users#set_password'
+  get '/forgot_password', to: 'forgot_passwords#forgot_password'
+  get '/confirm_password_reset', to: 'forgot_passwords#confirm_password_reset'
+  get '/new_password/:token', to: 'forgot_passwords#new_password'
+  post '/password_reset', to: 'forgot_passwords#password_reset'
+  post '/set_password', to: 'forgot_passwords#set_password'
 end
